put setup logic on another file
