+++ templates/index.html
@@ -17,19 +17,17 @@
 
     {{ util.flashed_messages(dismissible=True) }}
 
-    <div class="col-xs-6 col-md-4"></div>
-    <div class="col-xs-6 col-md-4">
-        <div class="row">
+    <div class="col-md-4 col-md-offset-4">
+        <div class="file-form-container">
             <h1>Upload a file:</h1>
             <form class="form" id="file_upload" method="POST" enctype=multipart/form-data action="{{ url_for('index') }}">
                 <div class="file-form form-group">
                     <label for="file_upload_field">File</label>
-                    <input class="" placeholder="Excel or CSV file" type="file" name="upload" id="file_upload_field">
+                    <input placeholder="Excel or CSV file" type="file" name="upload" id="file_upload_field">
                 </div>
-                <button type="submit" class="btn btn-primary">Submit</button>
+                <button type="submit" class="file-form-button btn btn-primary">Submit</button>
             </form>
         </div>
     </div>
-  <div class="col-xs-6 col-md-4"></div>
 </div>
 {% endblock %}
